move file browser files into folder
