+++ packages/editor/src/components/assets/FileBrowser/FileBrowserContentPanel.tsx
@@ -67,19 +67,19 @@ import { AssetLoader } from '@etherealengine/engine/src/assets/classes/AssetLoad
 import { SceneState } from '@etherealengine/engine/src/ecs/classes/Scene'
 import { archiverPath } from '@etherealengine/engine/src/schemas/media/archiver.schema'
 import { fileBrowserUploadPath } from '@etherealengine/engine/src/schemas/media/file-browser-upload.schema'
-import { SupportedFileTypes } from '../../constants/AssetTypes'
-import { inputFileWithAddToScene } from '../../functions/assetFunctions'
-import { bytesToSize, unique } from '../../functions/utils'
-import { Button } from '../inputs/Button'
-import StringInput from '../inputs/StringInput'
-import { ToolButton } from '../toolbar/ToolButton'
-import { AssetSelectionChangePropsType } from './AssetsPreviewPanel'
-import CompressionPanel from './CompressionPanel'
+import { SupportedFileTypes } from '../../../constants/AssetTypes'
+import { inputFileWithAddToScene } from '../../../functions/assetFunctions'
+import { bytesToSize, unique } from '../../../functions/utils'
+import { Button } from '../../inputs/Button'
+import StringInput from '../../inputs/StringInput'
+import { ToolButton } from '../../toolbar/ToolButton'
+import { AssetSelectionChangePropsType } from '../AssetsPreviewPanel'
+import CompressionPanel from '../CompressionPanel'
+import ImageConvertPanel from '../ImageConvertPanel'
+import styles from '../styles.module.scss'
 import { FileBrowserItem } from './FileBrowserGrid'
 import { FileDataType } from './FileDataType'
 import { FilePropertiesPanel } from './FilePropertiesPanel'
-import ImageConvertPanel from './ImageConvertPanel'
-import styles from './styles.module.scss'
 
 export const FileIconType = {
   gltf: ViewInArIcon,
